v. 3.0.2
pull request of team commits

+++ src/System/Ipv4Address.cpp
@@ -1,23 +1,9 @@
-// Copyright (c) 2011-2015 The Cryptonote developers
-// Copyright (c) 2015-2016 The Bytecoin developers
-// Copyright (c) 2016-2017 The TurtleCoin developers
-// Copyright (c) 2017-2018 krypt0x aka krypt0chaos
+// Copyright (c) 2011-2016 The Cryptonote developers
+// Copyright (c) 2016-2018 krypt0x aka krypt0chaos
 // Copyright (c) 2018 The Circle Foundation
 //
-// This file is part of Conceal Sense Crypto Engine.
-//
-// Conceal is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// Conceal is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with Conceal.  If not, see <http://www.gnu.org/licenses/>.
+// Distributed under the MIT/X11 software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "Ipv4Address.h"
 #include <stdexcept>
@@ -119,11 +105,11 @@ bool Ipv4Address::isLoopback() const {
 bool Ipv4Address::isPrivate() const {
   return
     // 10.0.0.0/8
-    (int)(value & 0xff000000) == (int)(10 << 24) ||
+    (value & 0xff000000) == (10 << 24) ||
     // 172.16.0.0/12
-    (int)(value & 0xfff00000) == (int)((172 << 24) | (16 << 16)) ||
+    (value & 0xfff00000) == ((172 << 24) | (16 << 16)) ||
     // 192.168.0.0/16
-    (int)(value & 0xffff0000) == (int)((192 << 24) | (168 << 16));
+    (value & 0xffff0000) == ((192 << 24) | (168 << 16));
 }
 
 }
